pyproject.toml
[project]
name = "industrial-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyqt6>=6.10.2",
    "python-snap7>=2.1.0",
    "websockets>=15.0",
]
